修正とパッケージ分け
